+++ src/config.ts
@@ -1,23 +1,19 @@
 export const DECIMALS: { [key: string]: number } = {
     EOS: 4,
-    PBTC: 8,
-    KGRAM: 4
+    PBTC: 8
 }
 
 export const POSITION_CONTRACTS: { [key: string]: string } = {
     EOS: "eosdtcntract",
-    PBTC: "eosdtpbtcpos",
-    KGRAM: "eosdtgrampos"
+    PBTC: "eosdtpbtcpos"
 }
 
 export const LIQUIDATOR_CONTRACTS: { [key: string]: string } = {
     EOS: "eosdtliqdatr",
-    PBTC: "eosdtpbtcliq",
-    KGRAM: "eosdtgramliq"
+    PBTC: "eosdtpbtcliq"
 }
 
 export const TOKEN_CONTRACTS: { [key: string]: string } = {
     EOS: "eosio.token",
-    PBTC: "btc.ptokens",
-    KGRAM: "eosdtkgtoken"
+    PBTC: "btc.ptokens"
 }